Added a config variable for database name called MONGODB_DB_NAME
In operation_primatives.py there was an assumption that the name of the mongo database was the same as the value of the APP_NAME environmental variable. I added the option of explicitly specifying the name of the database which is necessary if it's different than APP_NAME. This new environmental variable is called MONGODB_DB_NAME. If it's not specified than APP_NAME will be used.